Bringing the code up to date... Including:
Added NewMonthlyShowObject and NewWeeklyShowObject, based on
NewDailyShowObject - which has been updated to reflect the fact that we
should be returning the object itself if we're not returning false (this
probably needs to be done elsewhere as well!)

The API class returns a string containing which shows to generate when
running the runshows function, rather than returning the data which was
generated - which will make the external scripts a little easier to run!

ShowBroker:: has a new function - getShowsByIDs, which returns the
shows associated to a series of show IDs rather than just one showID.

ShowObject:: only contains the tracks associated to it, if requested.
Also, it now returns a spoken version of the show name when requested.

UI::getPronouncableDate now returns on yyyy, yyyymm and yyyymmdd rather
than just yyyymm and yyyymmdd.

VoteBroker::getVotesForTrackByShow now returns the adjusted votes again.
Don't know how I missed that one off!