Various changes
* Added a function to get 30 days of chart data for a track in
ChartBroker::
* Corrected HTML:: functions, ensuring that objects are no longer being
passed into the template for the Track() and Show() functions.
* Fixed a typo in ShowObject
* Made ShowTrackBroker::getShowTracksByTrackID() cache track
information.
* Added TrackBroker::getTotalTracks() function to count how many tracks
are across the whole system, and made it caching.
* TrackObject->getSelf() returns historical data about chart positions,
and it's own position in the chart (top40, top half, bottom half).
* Added some validation to the VoteBroker::getVotesForTrackByShow()
ensuring that existing votes for shows that didn't feature that track
don't show up. Also, capturing those votes as "votes for no show".
* Fixed some URLs in frontpage.html.tpl, show.html.tpl
* Wrote initial content for voted.html.tpl.
* Some formatting tweaks for voted.html.tpl and VoteObject