Various fixes and additions.
* Added an sources file to the externals directory (to show where to
find out more about the projects being used)
* Added two new libraries (jQuery.sparkline and PHPQRcode). Amended
.gitignore and libraries.json with this detail.
* Added the playlist.js into a new directory called /JAVASCRIPT which
contains local javascript functions not included from another library.
* ChartBroker::getLastThirtyDaysOfChartDataForOneTrack returns 30 days
worth of data, albeit with "null" for the days which have no entries in
the chart.
* The HTML class loads the sparkline library for all templates.
* ShowObject and TrackObject return QRCode PNG files for each item.
* TrackObject also returns lots more data about what position that track
was in the chart over the past 30 days, including highs and lows.
* UI class knows how to return a QR code object - if the file doesn't
already exist. Will be a bitch the first time it tries to load them all,
but after that, should be fine.
* Frontpage template doesn't load the inspector any more.
* Track template includes lots more new details.
* Missing function finder stops looking for QRcode classes.
* Index looks in the right place for media files.