Some bugfixes and some new code - all moving towards getting that all-important admin interface up and working!
* GenericObject::asBoolean actually returns a boolean (rather than a
guess)
* HTML class passes user details into Track and Show listings (so we
can add code into the templates to go to the admin interface for these
items if we're an admin). It also handles users better in the admin 
functions. Corrected a typo, moved a few lines of code around to make it
more legible. Has the 7 days timeout on the session, rather than 5
minutes (or whatever it is that the OpenID functions set it to!).
Corrected a typo, and made some redirections rather than loading
templates for inappropriate paths.
* Fixed a typo in the UserBroker and a redirection error in index
* Created rubbish admin and login templates.