Lots of changes, mostly around getting the showeditor to work, but with some typo fixing, and the removal of functions relating to the direct lookup of a daily show.
* Fixed php_codesniffer issues in API and HTML classes.
* GenericObject class now handles updates to tables which have no single
primary key (but instead have multiple)
* Fixed an object-interpretation error in HTML class. Also, now made the
show editor work for moving tracks up and down in the running order, and
removing tracks from a show.
* Fixed a Left Join SQL issue I'd been battling with in the
NewDailyShowObject class (had a moment of clarity!)
* The ShowObject now handles refreshes of the list of tracks.
* The functions to handle moving ShowTracks up, down and removing them
are now in the ShowTrackBroker.
* Added $arrDBKeyCol to the ShowTrackObject to handle updating a row
with no unique single unique column to sort by.
* Removed getTrackByDailyShowDate from TrackBroker as it's no longer
needed.
* Added TracksBroker::getTracksByShowIDOrderedByPartID for clarity.
* Removed code from the VoteBroker to look for DailyShow data, instead
of using the Show table.
* Added a new class in the CSS to handle changing an HTML element from
block to inline.
* Made the Show.html.tpl point the vote link to the vote link for that
show.
* Made showeditor.html.tpl more consistent with the "themeing" in the
rest of the site, also made it list the tracks in that show, and have
up, down, delete and edit buttons next to each track.
* Fixed a typo in the track_detail.tpl file.