Fixed the first item on the newly renamed todo_list.php (still a PHP file for the eclipse "Tasks" list), plus some other "noticed" bits.
* HTML class was using the raw chart data, rather than pulling just
the bits that were needed.
* TrackObject is now more efficient at not having two sets of variables
holding the same data about shows.
* Show.html.tpl links to the track via the QR code (probably needs
tidying up!)
* Created one include file for the show track vote links (leading to
consistency between the RSS feed and the HTML page). Due to some
template placement on the HTML page, had to use the variable "realshow"
rather than "show" in the include page, thus an assignment exists in the
two linked templates.
* track_detail.tpl only shows the shows table if the shows variables are
in the track detail. Tidied up some whitespace too.
* voted.html.tpl pulls in the tracks played in the rest of the show as
well as the details about the track you voted for, meaning if you're
listening to a show, you can vote as you go along, rathen than needing
to back out to the show listing and then vote again.